requirements.txt
Django
djangorestframework
django-cors
django-phonenumber-field[phonenumbers]
drf-yasg
django-filter
pandas